// SalesViewPCF/cds-generated/entities/EnvironmentVariableDefinition.ts
export const environmentvariabledefinitionMetadata = {
  typeName: "mscrm.environmentvariabledefinition",
  logicalName: "environmentvariabledefinition",
  collectionName: "environmentvariabledefinitions",
  primaryIdAttribute: "environmentvariabledefinitionid",
  attributeTypes: {
    // Numeric Types
    importsequencenumber: "Integer",
    iscustomizable: "ManagedProperty",
    timezoneruleversionnumber: "Integer",
    utcconversiontimezonecode: "Integer",
    versionnumber: "BigInt",
    // Optionsets
    componentstate: "Optionset",
    secretstore: "Optionset",
    statecode: "Optionset",
    statuscode: "Optionset",
    type: "Optionset",
    // Date Formats
    createdon: "DateAndTime:UserLocal",
    modifiedon: "DateAndTime:UserLocal",
    overriddencreatedon: "DateOnly:UserLocal",
    overwritetime: "DateOnly:UserLocal",
  },
  navigation: {
    owninguser: ["mscrm.systemuser"],
    owningteam: ["mscrm.team"],
    owningbusinessunit: ["mscrm.businessunit"],
    ownerid: ["mscrm.principal"],
    modifiedonbehalfby: ["mscrm.systemuser"],
    modifiedby: ["mscrm.systemuser"],
    createdonbehalfby: ["mscrm.systemuser"],
    createdby: ["mscrm.systemuser"],
    ParentDefinitionId: ["mscrm.environmentvariabledefinition"],
  },
};

// Attribute constants
export enum EnvironmentVariableDefinitionAttributes {
  ApiId = "apiid",
  ComponentState = "componentstate",
  ConnectionReferenceId = "connectionreferenceid",
  CreatedBy = "createdby",
  CreatedByName = "createdbyname",
  CreatedByYomiName = "createdbyyominame",
  CreatedOn = "createdon",
  CreatedOnBehalfBy = "createdonbehalfby",
  CreatedOnBehalfByName = "createdonbehalfbyname",
  CreatedOnBehalfByYomiName = "createdonbehalfbyyominame",
  DefaultValue = "defaultvalue",
  Description = "description",
  DisplayName = "displayname",
  EnvironmentVariableDefinitionId = "environmentvariabledefinitionid",
  EnvironmentVariableDefinitionIdUnique = "environmentvariabledefinitionidunique",
  Hint = "hint",
  ImportSequenceNumber = "importsequencenumber",
  IntroducedVersion = "introducedversion",
  IsCustomizable = "iscustomizable",
  IsManaged = "ismanaged",
  IsRequired = "isrequired",
  ModifiedBy = "modifiedby",
  ModifiedByName = "modifiedbyname",
  ModifiedByYomiName = "modifiedbyyominame",
  ModifiedOn = "modifiedon",
  ModifiedOnBehalfBy = "modifiedonbehalfby",
  ModifiedOnBehalfByName = "modifiedonbehalfbyname",
  ModifiedOnBehalfByYomiName = "modifiedonbehalfbyyominame",
  OverriddenCreatedOn = "overriddencreatedon",
  OverwriteTime = "overwritetime",
  OwnerId = "ownerid",
  OwnerIdName = "owneridname",
  OwnerIdType = "owneridtype",
  OwnerIdYomiName = "owneridyominame",
  OwningBusinessUnit = "owningbusinessunit",
  OwningBusinessUnitName = "owningbusinessunitname",
  OwningTeam = "owningteam",
  OwningUser = "owninguser",
  ParameterKey = "parameterkey",
  ParentDefinitionId = "parentdefinitionid",
  ParentDefinitionIdName = "parentdefinitionidname",
  SchemaName = "schemaname",
  SecretStore = "secretstore",
  SolutionId = "solutionid",
  statecode = "statecode",
  statuscode = "statuscode",
  SupportingSolutionId = "supportingsolutionid",
  TimeZoneRuleVersionNumber = "timezoneruleversionnumber",
  Type = "type",
  UTCConversionTimeZoneCode = "utcconversiontimezonecode",
  ValueSchema = "valueschema",
  VersionNumber = "versionnumber",
}