// SalesViewPCF/cds-generated/enums/axa_salesfulfillmentstatus_axa_salesfulfillmentstatus_axa_doescustomerhavedatagovernanceform.ts
/* eslint-disable*/
// Enum axa_salesfulfillmentstatus_axa_salesfulfillmentstatus_axa_doescustomerhavedatagovernanceform
export const enum axa_salesfulfillmentstatus_axa_salesfulfillmentstatus_axa_doescustomerhavedatagovernanceform {
  Pendingreview = 752580000,
  Yes = 752580001,
  No = 752580002,
}
  